server part modify
