+++ gpt.sql
@@ -9,19 +9,15 @@ CREATE TABLE `USER` (
   PRIMARY KEY (`id`)
 );
 
-
 CREATE TABLE `kakao` (
   `kakaoId` varchar(20) NOT NULL,
   `kakao_nickname` varchar(10) NOT NULL,
   PRIMARY KEY (`kakaoId`)
 );
 
-
-
-
 CREATE TABLE `ROOM` (
   `room_id` int NOT NULL AUTO_INCREMENT,
-  `id` varchar(255) NOT NULL,
+  `id` varchar(255) CHARACTER SET utf8mb3 COLLATE utf8mb3_general_ci DEFAULT NULL,
   `situation` varchar(50) NOT NULL,
   `accent` varchar(50) NOT NULL,
   `language` varchar(50) NOT NULL,
@@ -29,8 +25,8 @@ CREATE TABLE `ROOM` (
   PRIMARY KEY (`room_id`),
   KEY `id` (`id`),
   KEY `kakaoId` (`kakaoId`),
-  CONSTRAINT `room_ibfk_477` FOREIGN KEY (`id`) REFERENCES `USER` (`id`) ON DELETE SET NULL ON UPDATE CASCADE,
-  CONSTRAINT `room_ibfk_478` FOREIGN KEY (`kakaoId`) REFERENCES `kakao` (`kakaoId`) ON DELETE SET NULL ON UPDATE CASCADE
+  CONSTRAINT `room_ibfk_785` FOREIGN KEY (`id`) REFERENCES `USER` (`id`) ON DELETE SET NULL ON UPDATE CASCADE,
+  CONSTRAINT `room_ibfk_786` FOREIGN KEY (`kakaoId`) REFERENCES `kakao` (`kakaoId`) ON DELETE SET NULL ON UPDATE CASCADE
 );
 
 CREATE TABLE `MSG` (
@@ -46,15 +42,3 @@ CREATE TABLE `MSG` (
 
 
 
-
-
-create table `MSG`(
-    `msg_id` int NOT NULL AUTO_INCREMET,
-    `part_id` varchar(20) NOT NULL,
-    `content` text NOT NULL,
-    `room_id` int NOT NULL,
-    `user_id` varchar(225) NOT NULL,
-    PRIMARY KEY (`msg_id`),
-    KEY `room_id` (`room_id`),
-    CONSTRAINT `msg_ibfk_1` FOREIGN KEY (`room_id`) REFERENCES `ROOM` (`room_id`) ON DELETE CASCADE
-    );
